build ProfileCardContent.vue for DRY purpose / add it to ContributionItem.vue
